Change in data_period for CSV-LSTM
These change how the data_period variable works. Here, it is now an integer, and used in the CSV program. This makes it so the CSV file can have the maximum number of elements (good for versatility), while not sacrificing limiting the amount of time the model will look at.