Created Sync.py
This is used for updating the dataset with new dates. This is done instead of everything in the csv file.

I also added a 2d array. In the first column there is a date from the dataset. In the other column, there is an array with "gap" dates. I figured this would be useful in getting news stories that fall in between two dates in the dataset.